# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI短片工坊",
    "description": "本地单用户短片工坊：M1故事版本与DeepSeek适配器；真实文本验收及后续成片能力单独验证。",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/projects/{pid}/finishing": {
      "get": {
        "tags": [
          "finishing"
        ],
        "summary": "Get State",
        "operationId": "get_state_api_v1_projects__pid__finishing_get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "finishing"
        ],
        "summary": "Save",
        "operationId": "save_api_v1_projects__pid__finishing_put",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditSave"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{pid}/finishing/defaults": {
      "get": {
        "tags": [
          "finishing"
        ],
        "summary": "Defaults",
        "operationId": "defaults_api_v1_projects__pid__finishing_defaults_get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{pid}/exports": {
      "post": {
        "tags": [
          "finishing"
        ],
        "summary": "Submit",
        "operationId": "submit_api_v1_projects__pid__exports_post",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pid"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportCreate"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{pid}/exports/{jid}/retry": {
      "post": {
        "tags": [
          "finishing"
        ],
        "summary": "Retry",
        "operationId": "retry_api_v1_projects__pid__exports__jid__retry_post",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pid"
            }
          },
          {
            "name": "jid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Jid"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{pid}/exports/{jid}/download": {
      "get": {
        "tags": [
          "finishing"
        ],
        "summary": "Download",
        "operationId": "download_api_v1_projects__pid__exports__jid__download_get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pid"
            }
          },
          {
            "name": "jid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Jid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{pid}/finishing/music": {
      "post": {
        "tags": [
          "finishing"
        ],
        "summary": "Upload Music",
        "operationId": "upload_music_api_v1_projects__pid__finishing_music_post",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_music_api_v1_projects__pid__finishing_music_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{pid}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project_api_v1_projects__pid__get",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Edit Project",
        "operationId": "edit_project_api_v1_projects__pid__patch",
        "parameters": [
          {
            "name": "pid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Pid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_music_api_v1_projects__pid__finishing_music_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_music_api_v1_projects__pid__finishing_music_post"
      },
      "Clip": {
        "properties": {
          "shot_id": {
            "type": "string",
            "format": "uuid",
            "title": "Shot Id"
          },
          "video_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Id"
          },
          "trim_start": {
            "type": "number",
            "maximum": 36000.0,
            "minimum": 0.0,
            "title": "Trim Start",
            "default": 0
          },
          "duration": {
            "type": "number",
            "maximum": 600.0,
            "exclusiveMinimum": 0.0,
            "title": "Duration"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/LineTiming"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Lines"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "shot_id",
          "duration"
        ],
        "title": "Clip"
      },
...
      "EditDraft": {
        "properties": {
          "board_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Board Version Id"
          },
          "filename": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "pattern": "^[^/\\\\\\x00-\\x1f]+$",
            "title": "Filename",
            "default": "我的作品"
          },
          "clips": {
            "items": {
              "$ref": "#/components/schemas/Clip"
            },
            "type": "array",
            "maxItems": 200,
            "title": "Clips"
          },
          "fps": {
            "type": "integer",
            "enum": [
              24,
              25,
              30
            ],
            "title": "Fps",
            "default": 24
          },
          "fit": {
            "type": "string",
            "enum": [
              "pad",
              "crop"
            ],
            "title": "Fit",
            "default": "pad"
          },
          "narration": {
            "type": "boolean",
            "title": "Narration",
            "default": true
          },
          "subtitles": {
            "type": "boolean",
            "title": "Subtitles",
            "default": true
          },
          "original_audio": {
            "type": "boolean",
            "title": "Original Audio",
            "default": false
          },
          "voice_volume": {
            "type": "number",
            "maximum": 2.0,
            "minimum": 0.1,
            "title": "Voice Volume",
            "default": 1
          },
          "original_volume": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Original Volume",
            "default": 0.5
          },
          "music_file_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Music File Id"
          },
          "music_volume": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Music Volume",
            "default": 0.15
          },
          "continuity_ack": {
            "type": "boolean",
            "title": "Continuity Ack",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "EditDraft"
      },
      "EditSave": {
        "properties": {
          "revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Revision"
          },
          "draft": {
            "$ref": "#/components/schemas/EditDraft"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision",
          "draft"
        ],
        "title": "EditSave"
      },
      "EditState": {
        "properties": {
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "draft": {
            "$ref": "#/components/schemas/EditDraft"
          },
          "specification": {
            "additionalProperties": true,
            "type": "object",
            "title": "Specification"
          },
          "blockers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blockers"
          },
          "timeline": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Timeline"
          },
          "exports": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Exports"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision",
          "draft",
          "specification",
          "blockers",
          "timeline",
          "exports"
        ],
        "title": "EditState"
      },
...
      "ExportCreate": {
        "properties": {
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision"
        ],
        "title": "ExportCreate"
      },
      "FileOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "mime": {
            "type": "string",
            "title": "Mime"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "project_id",
          "filename",
          "mime",
          "size",
          "sha256",
          "created_at"
        ],
        "title": "FileOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "snapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Snapshot"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "project_id",
          "kind",
          "state",
          "snapshot",
          "error",
          "created_at",
          "updated_at"
        ],
        "title": "JobOut"
      },
...
      "LineTiming": {
        "properties": {
          "line_id": {
            "type": "string",
            "format": "uuid",
            "title": "Line Id"
          },
          "start": {
            "type": "number",
            "maximum": 36000.0,
            "minimum": 0.0,
            "title": "Start"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "line_id",
          "start"
        ],
        "title": "LineTiming"
      },
...
      "ProjectOut": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Name"
          },
          "market": {
            "type": "string",
            "enum": [
              "zh",
              "en"
            ],
            "title": "Market",
            "default": "zh"
          },
          "type_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type Id"
          },
          "aspect_ratio": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "9:16",
                  "16:9",
                  "1:1"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Aspect Ratio"
          },
          "resolution": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "720P",
                  "1080P",
                  "4K"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          },
          "style_resource_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Style Resource Id"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "generation_settings": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Settings"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "id",
          "revision",
          "stage",
          "status",
          "generation_settings",
          "updated_at"
        ],
        "title": "ProjectOut"
      },
      "ProjectPatch": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Name"
          },
          "revision": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "revision"
        ],
        "title": "ProjectPatch"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}